+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 * Viewer: Minor rendering improvements
 * Viewer: Switched to CompiledBindings for improved performance
 * Viewer: Updated New Game Window
-* Viewer: Updated Avalonia to 11.0.0-preview5
-* Viewer: Updated Markdown.Avalonia to 11.0.0-a9
+* Viewer: Updated Avalonia to 11.0.0-preview6
+* Viewer: Updated Markdown.Avalonia to Markdown.Avalonia.Tight 11.0.0-b1
 
 ## v0.13.2 ##